extend VisSelectedTimeClippedModule to VisSelectedTimeWarningModule
