+++ tensorboard/webapp/metrics/views/card_renderer/BUILD
@@ -95,7 +95,7 @@ tf_ng_module(
     deps = [
         ":run_name",
         ":utils",
-        ":vis_selected_time_clipped",
+        ":vis_selected_time_warning",
         "//tensorboard/webapp:app_state",
         "//tensorboard/webapp/angular:expect_angular_material_button",
         "//tensorboard/webapp/angular:expect_angular_material_icon",
@@ -191,19 +191,19 @@ tf_ng_module(
 )
 
 tf_sass_binary(
-    name = "vis_selected_time_clipped_styles",
-    src = "vis_selected_time_clipped_component.scss",
+    name = "vis_selected_time_warning_styles",
+    src = "vis_selected_time_warning_component.scss",
     deps = ["//tensorboard/webapp:angular_material_sass_deps"],
 )
 
 tf_ng_module(
-    name = "vis_selected_time_clipped",
+    name = "vis_selected_time_warning",
     srcs = [
-        "vis_selected_time_clipped_component.ts",
-        "vis_selected_time_clipped_module.ts",
+        "vis_selected_time_warning_component.ts",
+        "vis_selected_time_warning_module.ts",
     ],
     assets = [
-        ":vis_selected_time_clipped_styles",
+        ":vis_selected_time_warning_styles",
     ],
     deps = [
         "//tensorboard/webapp/angular:expect_angular_material_icon",
@@ -260,7 +260,7 @@ tf_ng_module(
         ":data_download_dialog",
         ":scalar_card_types",
         ":utils",
-        ":vis_selected_time_clipped",
+        ":vis_selected_time_warning",
         "//tensorboard/webapp:app_state",
         "//tensorboard/webapp:selectors",
         "//tensorboard/webapp/angular:expect_angular_cdk_overlay",
@@ -318,7 +318,7 @@ tf_ts_library(
         ":scalar_card",
         ":scalar_card_types",
         ":utils",
-        ":vis_selected_time_clipped",
+        ":vis_selected_time_warning",
         "//tensorboard/webapp:app_state",
         "//tensorboard/webapp:selectors",
         "//tensorboard/webapp/angular:expect_angular_cdk_overlay",
