+++ tensorboard/webapp/metrics/views/card_renderer/histogram_card_module.ts
@@ -22,7 +22,7 @@ import {TruncatedPathModule} from '../../../widgets/text/truncated_path_module';
 import {HistogramCardComponent} from './histogram_card_component';
 import {HistogramCardContainer} from './histogram_card_container';
 import {RunNameModule} from './run_name_module';
-import {VisSelectedTimeClippedModule} from './vis_selected_time_clipped_module';
+import {VisSelectedTimeWarningModule} from './vis_selected_time_warning_module';
 
 @NgModule({
   declarations: [HistogramCardContainer, HistogramCardComponent],
@@ -35,7 +35,7 @@ import {VisSelectedTimeClippedModule} from './vis_selected_time_clipped_module';
     MatProgressSpinnerModule,
     RunNameModule,
     TruncatedPathModule,
-    VisSelectedTimeClippedModule,
+    VisSelectedTimeWarningModule,
   ],
 })
 export class HistogramCardModule {}
