--- tensorboard/webapp/metrics/views/card_renderer/vis_selected_time_clipped_component.scss
@@ -1,32 +1,0 @@
-/* Copyright 2021 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-@use '@angular/material' as mat;
-@import 'tensorboard/webapp/theme/tb_theme';
-
-:host {
-  color: mat.get-color-from-palette(map-get($tb-theme, warn), 700);
-  height: 1em;
-  line-height: 0;
-  width: 1em;
-
-  @include tb-dark-theme {
-    color: mat.get-color-from-palette(map-get($tb-dark-theme, warn), 700);
-  }
-
-  mat-icon {
-    height: 100%;
-    width: 100%;
-  }
-}
